pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finbridge"
version = "0.1.0"
description = "Sync Saxo Bank portfolio to Google Sheets"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=5.3.3",
    "certifi>=2024.2.2",
    "charset-normalizer>=3.3.2",
    "click>=8.1.7",
    "flask>=3.0.3",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "gspread>=6.1.0",
    "idna>=3.7",
    "importlib-metadata>=8.2.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.4",
    "markupsafe>=2.1.5",
    "oauthlib>=3.2.2",
    "openpyxl",
    "pandas",
    "pyasn1>=0.6.0",
    "pyasn1-modules>=0.4.0",
    "requests>=2.31.0",
    "requests-oauthlib>=2.0.0",
    "rsa>=4.9",
    "saxo-openapi>=0.6.0",
    "strenum>=0.4.15",
    "tabula-py",
    "urllib3>=2.2.1",
    "werkzeug>=3.0.3",
    "zipp>=3.19.2",
]

[project.scripts]
finbridge = "finbridge.main:main"
finbridge-export = "finbridge.export_tables:main"

[tool.hatch.build.targets.wheel]
packages = ["src/finbridge"]
